requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
black==24.4.2
click==8.1.7
Django==5.0.6
django-cacheops==7.0.2
django-filter==24.2
django-redis==5.4.0
djangorestframework==3.15.1
factory-boy==3.3.0
Faker==25.8.0
funcy==2.0
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
psycopg2-binary==2.9.9
pytest==8.2.2
pytest-django==4.8.0
python-dateutil==2.9.0.post0
redis==5.0.5
six==1.16.0
sqlparse==0.5.0
